fixed all cached system
