--- agent-catalog/app/agent/[id]/layout.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-import { QueryProvider } from "../../ClientQueryProvider";
-
-export default function AgentLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <QueryProvider>{children}</QueryProvider>;
-}
