+++ agent-catalog/features/diagnosed/visionDiagnosis.ts
@@ -7,6 +7,9 @@ const additionalUrls = {
   vision_ollama: "/api/chat",
 };
 
+const TIME_OUT = 20000;
+const MIN_DURATION = 5000;
+
 export async function loadImage(
     url: string,
     maxWidth = 320,
@@ -64,12 +67,16 @@ export async function loadImage(
 async function safeFetch(
   fullUrl: string,
   options?: RequestInit,
+  timeoutMs = TIME_OUT
 ): Promise<EvaluationResult> {
+  const controller = new AbortController();
+  const id = setTimeout(() => controller.abort(), timeoutMs);
   const start = performance.now();
   try {
     if (!options) {
         const res = await fetch(fullUrl);
         const end = performance.now();
+        clearTimeout(id);
         const duration = end - start;
         const status = res.ok ? "pass" : "fail";
         const score = calculateScore({ status, duration });
@@ -78,16 +85,19 @@ async function safeFetch(
     } else {
         const res = await fetch(fullUrl, options);
         const end = performance.now();
+        clearTimeout(id);
         const duration = end - start;
         const status = res.ok ? "pass" : "fail";
         const score = calculateScore({ status, duration });
 
         return { status, score };
     }
-  } catch {
+  } catch (err: any) {
     const end = performance.now();
+    clearTimeout(id);
     const duration = end - start;
-    return { status: "fail", score: calculateScore({ status: "fail", duration }) };
+    const isAbort = err.name === "AbortError";
+    return { status: "fail", score: calculateScore({ status: "fail", duration, timeout: isAbort }) };
   }
 }
 
@@ -95,13 +105,16 @@ async function safeFetch(
 function calculateScore({
   status,
   duration,
+  timeout = false,
 }: {
   status: "pass" | "fail";
   duration: number;
+  timeout?: boolean;
 }): number {
+  if (timeout) return 0;
   let score = 0;
   if (status === "pass") score += 50;
-  if (duration < 5000) score += 50 * ((5000 - duration) / 5000); 
+  if (duration < MIN_DURATION) score += 50 * ((MIN_DURATION - duration) / MIN_DURATION); 
   return Math.round(score);
 }
